Convert a data frame to tree format

Convert a data frame to tree format

Usage

as_tree(x, ...)

Arguments

x

The data frame to convert.

id_columns

The columns to use to split the data frame into a tree, as a character vector. Should be ordered from least specific (upper branches) to most (lowest branches).

root_name

Optionally, the name of the root node.

drop_id_columns

Logical: Should the columns used to split the data frame into nodes be included as values in each node?

Examples

data(GNI2014, package = "treemap")
as_tree(GNI2014, id_columns = c("continent"), root_name = "world")
#> <dendro::tree>
#>  @ size        : int 9
#>  @ node_indices: Named int [1:9] 1 2 3 4 5 6 7 8 9
#>  .. - attr(*, "names")= chr [1:9] "world" "Africa" "Antarctica" "Asia" ...
#>  @ next_idx    : int 10
#>  @ root        : Named int 1
#>  .. - attr(*, "names")= chr "world"
#>  @ nodes       :List of 9
#>  .. $ world                  : <dendro::node>
#>  ..  ..@ index   : int 1
#>  ..  ..@ value   :'data.frame':	0 obs. of  0 variables
#>  ..  ..@ parent  : int NA
#>  ..  ..@ children: Named int [1:8] 2 3 4 5 6 7 8 9
#>  .. .. .. - attr(*, "names")= chr [1:8] "Africa" "Antarctica" "Asia" "Europe" ...
#>  .. $ Africa                 : <dendro::node>
#>  ..  ..@ index   : int 2
#>  ..  ..@ value   :'data.frame':	48 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:48] "GNQ" "GAB" "LBY" "BWA" ...
#>  .. .. .. $ country   : chr [1:48] "Equatorial Guinea" "Gabon" "Libya" "Botswana" ...
#>  .. .. .. $ population: num [1:48] 650702 1514993 6310434 1990876 49052489 ...
#>  .. .. .. $ GNI       : int [1:48] 10210 9720 7820 7240 6800 5630 5490 4230 3550 3450 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Antarctica             : <dendro::node>
#>  ..  ..@ index   : int 3
#>  ..  ..@ value   :'data.frame':	0 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr(0) 
#>  .. .. .. $ country   : chr(0) 
#>  .. .. .. $ population: num(0) 
#>  .. .. .. $ GNI       : int(0) 
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Asia                   : <dendro::node>
#>  ..  ..@ index   : int 4
#>  ..  ..@ value   :'data.frame':	47 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:47] "QAT" "MAC" "SGP" "KWT" ...
#>  .. .. .. $ country   : chr [1:47] "Qatar" "Macao SAR, China" "Singapore" "Kuwait" ...
#>  .. .. .. $ population: num [1:47] 833285 559846 4657542 2691158 4798491 ...
#>  .. .. .. $ GNI       : int [1:47] 92200 76270 55150 49300 44600 42000 40320 37320 35320 27090 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Europe                 : <dendro::node>
#>  ..  ..@ index   : int 5
#>  ..  ..@ value   :'data.frame':	41 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:41] "NOR" "CHE" "LUX" "SWE" ...
#>  .. .. .. $ country   : chr [1:41] "Norway" "Switzerland" "Luxembourg" "Sweden" ...
#>  .. .. .. $ population: num [1:41] 4676305 7604467 491775 9059651 5500510 ...
#>  .. .. .. $ GNI       : int [1:41] 103630 88120 75990 61610 61310 51890 49670 48420 47640 46350 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ North America          : <dendro::node>
#>  ..  ..@ index   : int 6
#>  ..  ..@ value   :'data.frame':	24 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:24] "BMU" "USA" "CAN" "BHS" ...
#>  .. .. .. $ country   : chr [1:24] "Bermuda" "United States" "Canada" "Bahamas, The" ...
#>  .. .. .. $ population: num [1:24] 6.78e+04 3.14e+08 3.35e+07 3.09e+05 1.31e+06 ...
#>  .. .. .. $ GNI       : int [1:24] 106140 55200 51630 20980 20070 19310 15310 14920 13300 11130 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Oceania                : <dendro::node>
#>  ..  ..@ index   : int 7
#>  ..  ..@ value   :'data.frame':	13 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:13] "AUS" "NZL" "PLW" "TUV" ...
#>  .. .. .. $ country   : chr [1:13] "Australia" "New Zealand" "Palau" "Tuvalu" ...
#>  .. .. .. $ population: num [1:13] 21262641 4213418 20796 10640 944720 ...
#>  .. .. .. $ GNI       : int [1:13] 64540 41070 11110 5720 4870 4390 4260 4060 3200 3160 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Seven seas (open ocean): <dendro::node>
#>  ..  ..@ index   : int 8
#>  ..  ..@ value   :'data.frame':	3 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:3] "SYC" "MUS" "MDV"
#>  .. .. .. $ country   : chr [1:3] "Seychelles" "Mauritius" "Maldives"
#>  .. .. .. $ population: num [1:3] 87476 1284264 396334
#>  .. .. .. $ GNI       : int [1:3] 14100 9630 6410
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ South America          : <dendro::node>
#>  ..  ..@ index   : int 9
#>  ..  ..@ value   :'data.frame':	12 obs. of  4 variables:
#>  .. .. .. $ iso3      : chr [1:12] "URY" "CHL" "ARG" "VEN" ...
#>  .. .. .. $ country   : chr [1:12] "Uruguay" "Chile" "Argentina" "Venezuela, RB" ...
#>  .. .. .. $ population: num [1:12] 3.49e+06 1.66e+07 4.09e+07 2.68e+07 1.99e+08 ...
#>  .. .. .. $ GNI       : int [1:12] 16350 14910 13480 12500 11530 9950 7970 6360 6090 4400 ...
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
as_tree(GNI2014, id_columns = c("continent", "country"), root_name = "world")
#> <dendro::tree>
#>  @ size        : int 197
#>  @ node_indices: Named int [1:197] 1 2 3 4 5 6 7 8 9 10 ...
#>  .. - attr(*, "names")= chr [1:197] "world" "Africa" "Algeria" "Benin" ...
#>  @ next_idx    : int 198
#>  @ root        : Named int 1
#>  .. - attr(*, "names")= chr "world"
#>  @ nodes       :List of 197
#>  .. $ world                         : <dendro::node>
#>  ..  ..@ index   : int 1
#>  ..  ..@ value   :'data.frame':	0 obs. of  0 variables
#>  ..  ..@ parent  : int NA
#>  ..  ..@ children: Named int [1:8] 2 51 52 100 142 167 181 185
#>  .. .. .. - attr(*, "names")= chr [1:8] "Africa" "Antarctica" "Asia" "Europe" ...
#>  .. $ Africa                        : <dendro::node>
#>  ..  ..@ index   : int 2
#>  ..  ..@ value   :'data.frame':	0 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr(0) 
#>  .. .. .. $ population: num(0) 
#>  .. .. .. $ GNI       : int(0) 
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: Named int [1:48] 3 4 5 6 7 8 9 10 11 12 ...
#>  .. .. .. - attr(*, "names")= chr [1:48] "Algeria" "Benin" "Botswana" "Burkina Faso" ...
#>  .. $ Algeria                       : <dendro::node>
#>  ..  ..@ index   : int 3
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "DZA"
#>  .. .. .. $ population: num 34178188
#>  .. .. .. $ GNI       : int 5490
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Benin                         : <dendro::node>
#>  ..  ..@ index   : int 4
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BEN"
#>  .. .. .. $ population: num 8791832
#>  .. .. .. $ GNI       : int 890
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Botswana                      : <dendro::node>
#>  ..  ..@ index   : int 5
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BWA"
#>  .. .. .. $ population: num 1990876
#>  .. .. .. $ GNI       : int 7240
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Burkina Faso                  : <dendro::node>
#>  ..  ..@ index   : int 6
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BFA"
#>  .. .. .. $ population: num 15746232
#>  .. .. .. $ GNI       : int 700
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Burundi                       : <dendro::node>
#>  ..  ..@ index   : int 7
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BDI"
#>  .. .. .. $ population: num 8988091
#>  .. .. .. $ GNI       : int 270
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Cabo Verde                    : <dendro::node>
#>  ..  ..@ index   : int 8
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CPV"
#>  .. .. .. $ population: num 429474
#>  .. .. .. $ GNI       : int 3450
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Cameroon                      : <dendro::node>
#>  ..  ..@ index   : int 9
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CMR"
#>  .. .. .. $ population: num 18879301
#>  .. .. .. $ GNI       : int 1350
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Central African Republic      : <dendro::node>
#>  ..  ..@ index   : int 10
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CAF"
#>  .. .. .. $ population: num 4511488
#>  .. .. .. $ GNI       : int 320
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Chad                          : <dendro::node>
#>  ..  ..@ index   : int 11
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TCD"
#>  .. .. .. $ population: num 10329208
#>  .. .. .. $ GNI       : int 980
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Comoros                       : <dendro::node>
#>  ..  ..@ index   : int 12
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "COM"
#>  .. .. .. $ population: num 752438
#>  .. .. .. $ GNI       : int 790
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Congo, Dem. Rep.              : <dendro::node>
#>  ..  ..@ index   : int 13
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "COD"
#>  .. .. .. $ population: num 68692542
#>  .. .. .. $ GNI       : int 380
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Congo, Rep.                   : <dendro::node>
#>  ..  ..@ index   : int 14
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "COG"
#>  .. .. .. $ population: num 4012809
#>  .. .. .. $ GNI       : int 2720
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Cote d'Ivoire                 : <dendro::node>
#>  ..  ..@ index   : int 15
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CIV"
#>  .. .. .. $ population: num 20617068
#>  .. .. .. $ GNI       : int 1450
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Egypt, Arab Rep.              : <dendro::node>
#>  ..  ..@ index   : int 16
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "EGY"
#>  .. .. .. $ population: num 83082869
#>  .. .. .. $ GNI       : int 3050
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Equatorial Guinea             : <dendro::node>
#>  ..  ..@ index   : int 17
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GNQ"
#>  .. .. .. $ population: num 650702
#>  .. .. .. $ GNI       : int 10210
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Ethiopia                      : <dendro::node>
#>  ..  ..@ index   : int 18
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ETH"
#>  .. .. .. $ population: num 85237338
#>  .. .. .. $ GNI       : int 550
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Gabon                         : <dendro::node>
#>  ..  ..@ index   : int 19
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GAB"
#>  .. .. .. $ population: num 1514993
#>  .. .. .. $ GNI       : int 9720
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Gambia, The                   : <dendro::node>
#>  ..  ..@ index   : int 20
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GMB"
#>  .. .. .. $ population: num 1782893
#>  .. .. .. $ GNI       : int 500
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Ghana                         : <dendro::node>
#>  ..  ..@ index   : int 21
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GHA"
#>  .. .. .. $ population: num 23832495
#>  .. .. .. $ GNI       : int 1590
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Guinea                        : <dendro::node>
#>  ..  ..@ index   : int 22
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GIN"
#>  .. .. .. $ population: num 10057975
#>  .. .. .. $ GNI       : int 470
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Guinea-Bissau                 : <dendro::node>
#>  ..  ..@ index   : int 23
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GNB"
#>  .. .. .. $ population: num 1533964
#>  .. .. .. $ GNI       : int 550
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Kenya                         : <dendro::node>
#>  ..  ..@ index   : int 24
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KEN"
#>  .. .. .. $ population: num 3.9e+07
#>  .. .. .. $ GNI       : int 1290
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Lesotho                       : <dendro::node>
#>  ..  ..@ index   : int 25
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LSO"
#>  .. .. .. $ population: num 2130819
#>  .. .. .. $ GNI       : int 1330
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Liberia                       : <dendro::node>
#>  ..  ..@ index   : int 26
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LBR"
#>  .. .. .. $ population: num 3441790
#>  .. .. .. $ GNI       : int 370
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Libya                         : <dendro::node>
#>  ..  ..@ index   : int 27
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LBY"
#>  .. .. .. $ population: num 6310434
#>  .. .. .. $ GNI       : int 7820
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Madagascar                    : <dendro::node>
#>  ..  ..@ index   : int 28
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MDG"
#>  .. .. .. $ population: num 20653556
#>  .. .. .. $ GNI       : int 440
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Malawi                        : <dendro::node>
#>  ..  ..@ index   : int 29
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MWI"
#>  .. .. .. $ population: num 14268711
#>  .. .. .. $ GNI       : int 250
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Mali                          : <dendro::node>
#>  ..  ..@ index   : int 30
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MLI"
#>  .. .. .. $ population: num 12666987
#>  .. .. .. $ GNI       : int 650
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Mauritania                    : <dendro::node>
#>  ..  ..@ index   : int 31
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MRT"
#>  .. .. .. $ population: num 3129486
#>  .. .. .. $ GNI       : int 1270
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Morocco                       : <dendro::node>
#>  ..  ..@ index   : int 32
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MAR"
#>  .. .. .. $ population: num 34859364
#>  .. .. .. $ GNI       : int 3070
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Mozambique                    : <dendro::node>
#>  ..  ..@ index   : int 33
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MOZ"
#>  .. .. .. $ population: num 21669278
#>  .. .. .. $ GNI       : int 600
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Namibia                       : <dendro::node>
#>  ..  ..@ index   : int 34
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "NAM"
#>  .. .. .. $ population: num 2108665
#>  .. .. .. $ GNI       : int 5630
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Niger                         : <dendro::node>
#>  ..  ..@ index   : int 35
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "NER"
#>  .. .. .. $ population: num 15306252
#>  .. .. .. $ GNI       : int 410
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Nigeria                       : <dendro::node>
#>  ..  ..@ index   : int 36
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "NGA"
#>  .. .. .. $ population: num 1.49e+08
#>  .. .. .. $ GNI       : int 2970
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Rwanda                        : <dendro::node>
#>  ..  ..@ index   : int 37
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "RWA"
#>  .. .. .. $ population: num 10473282
#>  .. .. .. $ GNI       : int 700
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Sao Tome and Principe         : <dendro::node>
#>  ..  ..@ index   : int 38
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "STP"
#>  .. .. .. $ population: num 212679
#>  .. .. .. $ GNI       : int 1670
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Senegal                       : <dendro::node>
#>  ..  ..@ index   : int 39
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SEN"
#>  .. .. .. $ population: num 13711597
#>  .. .. .. $ GNI       : int 1050
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Sierra Leone                  : <dendro::node>
#>  ..  ..@ index   : int 40
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SLE"
#>  .. .. .. $ population: num 6440053
#>  .. .. .. $ GNI       : int 700
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ South Africa                  : <dendro::node>
#>  ..  ..@ index   : int 41
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ZAF"
#>  .. .. .. $ population: num 49052489
#>  .. .. .. $ GNI       : int 6800
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ South Sudan                   : <dendro::node>
#>  ..  ..@ index   : int 42
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SSD"
#>  .. .. .. $ population: num 10625176
#>  .. .. .. $ GNI       : int 970
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Sudan                         : <dendro::node>
#>  ..  ..@ index   : int 43
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SDN"
#>  .. .. .. $ population: num 25946220
#>  .. .. .. $ GNI       : int 1710
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Swaziland                     : <dendro::node>
#>  ..  ..@ index   : int 44
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SWZ"
#>  .. .. .. $ population: num 1123913
#>  .. .. .. $ GNI       : int 3550
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Tanzania                      : <dendro::node>
#>  ..  ..@ index   : int 45
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TZA"
#>  .. .. .. $ population: num 4.1e+07
#>  .. .. .. $ GNI       : int 920
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Togo                          : <dendro::node>
#>  ..  ..@ index   : int 46
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TGO"
#>  .. .. .. $ population: num 6019877
#>  .. .. .. $ GNI       : int 570
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Tunisia                       : <dendro::node>
#>  ..  ..@ index   : int 47
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TUN"
#>  .. .. .. $ population: num 10486339
#>  .. .. .. $ GNI       : int 4230
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Uganda                        : <dendro::node>
#>  ..  ..@ index   : int 48
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "UGA"
#>  .. .. .. $ population: num 32369558
#>  .. .. .. $ GNI       : int 670
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Zambia                        : <dendro::node>
#>  ..  ..@ index   : int 49
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ZMB"
#>  .. .. .. $ population: num 11862740
#>  .. .. .. $ GNI       : int 1680
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Zimbabwe                      : <dendro::node>
#>  ..  ..@ index   : int 50
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ZWE"
#>  .. .. .. $ population: num 12619600
#>  .. .. .. $ GNI       : int 840
#>  ..  ..@ parent  : int 2
#>  ..  ..@ children: int(0) 
#>  .. $ Antarctica                    : <dendro::node>
#>  ..  ..@ index   : int 51
#>  ..  ..@ value   :'data.frame':	0 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr(0) 
#>  .. .. .. $ population: num(0) 
#>  .. .. .. $ GNI       : int(0) 
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: int(0) 
#>  .. $ Asia                          : <dendro::node>
#>  ..  ..@ index   : int 52
#>  ..  ..@ value   :'data.frame':	0 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr(0) 
#>  .. .. .. $ population: num(0) 
#>  .. .. .. $ GNI       : int(0) 
#>  ..  ..@ parent  : int 1
#>  ..  ..@ children: Named int [1:47] 53 54 55 56 57 58 59 60 61 62 ...
#>  .. .. .. - attr(*, "names")= chr [1:47] "Afghanistan" "Armenia" "Azerbaijan" "Bahrain" ...
#>  .. $ Afghanistan                   : <dendro::node>
#>  ..  ..@ index   : int 53
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "AFG"
#>  .. .. .. $ population: num 28400000
#>  .. .. .. $ GNI       : int 680
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Armenia                       : <dendro::node>
#>  ..  ..@ index   : int 54
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ARM"
#>  .. .. .. $ population: num 2967004
#>  .. .. .. $ GNI       : int 4020
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Azerbaijan                    : <dendro::node>
#>  ..  ..@ index   : int 55
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "AZE"
#>  .. .. .. $ population: num 8238672
#>  .. .. .. $ GNI       : int 7590
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Bahrain                       : <dendro::node>
#>  ..  ..@ index   : int 56
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BHR"
#>  .. .. .. $ population: num 727785
#>  .. .. .. $ GNI       : int 21060
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Bangladesh                    : <dendro::node>
#>  ..  ..@ index   : int 57
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BGD"
#>  .. .. .. $ population: num 1.56e+08
#>  .. .. .. $ GNI       : int 1080
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Bhutan                        : <dendro::node>
#>  ..  ..@ index   : int 58
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BTN"
#>  .. .. .. $ population: num 691141
#>  .. .. .. $ GNI       : int 2370
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Brunei Darussalam             : <dendro::node>
#>  ..  ..@ index   : int 59
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "BRN"
#>  .. .. .. $ population: num 388190
#>  .. .. .. $ GNI       : int 37320
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Cambodia                      : <dendro::node>
#>  ..  ..@ index   : int 60
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KHM"
#>  .. .. .. $ population: num 14494293
#>  .. .. .. $ GNI       : int 1020
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ China                         : <dendro::node>
#>  ..  ..@ index   : int 61
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CHN"
#>  .. .. .. $ population: num 1.34e+09
#>  .. .. .. $ GNI       : int 7400
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Cyprus                        : <dendro::node>
#>  ..  ..@ index   : int 62
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "CYP"
#>  .. .. .. $ population: num 531640
#>  .. .. .. $ GNI       : int 26370
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Georgia                       : <dendro::node>
#>  ..  ..@ index   : int 63
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "GEO"
#>  .. .. .. $ population: num 4615807
#>  .. .. .. $ GNI       : int 3720
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Hong Kong SAR, China          : <dendro::node>
#>  ..  ..@ index   : int 64
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "HKG"
#>  .. .. .. $ population: num 7061200
#>  .. .. .. $ GNI       : int 40320
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ India                         : <dendro::node>
#>  ..  ..@ index   : int 65
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "IND"
#>  .. .. .. $ population: num 1.17e+09
#>  .. .. .. $ GNI       : int 1570
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Indonesia                     : <dendro::node>
#>  ..  ..@ index   : int 66
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "IDN"
#>  .. .. .. $ population: num 2.4e+08
#>  .. .. .. $ GNI       : int 3630
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Iran, Islamic Rep.            : <dendro::node>
#>  ..  ..@ index   : int 67
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "IRN"
#>  .. .. .. $ population: num 66429284
#>  .. .. .. $ GNI       : int 7120
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Iraq                          : <dendro::node>
#>  ..  ..@ index   : int 68
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "IRQ"
#>  .. .. .. $ population: num 31129225
#>  .. .. .. $ GNI       : int 6500
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Israel                        : <dendro::node>
#>  ..  ..@ index   : int 69
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ISR"
#>  .. .. .. $ population: num 7233701
#>  .. .. .. $ GNI       : int 35320
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Japan                         : <dendro::node>
#>  ..  ..@ index   : int 70
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "JPN"
#>  .. .. .. $ population: num 1.27e+08
#>  .. .. .. $ GNI       : int 42000
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Jordan                        : <dendro::node>
#>  ..  ..@ index   : int 71
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "JOR"
#>  .. .. .. $ population: num 6342948
#>  .. .. .. $ GNI       : int 5160
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Kazakhstan                    : <dendro::node>
#>  ..  ..@ index   : int 72
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KAZ"
#>  .. .. .. $ population: num 15399437
#>  .. .. .. $ GNI       : int 11850
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Korea, Rep.                   : <dendro::node>
#>  ..  ..@ index   : int 73
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KOR"
#>  .. .. .. $ population: num 48508972
#>  .. .. .. $ GNI       : int 27090
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Kuwait                        : <dendro::node>
#>  ..  ..@ index   : int 74
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KWT"
#>  .. .. .. $ population: num 2691158
#>  .. .. .. $ GNI       : int 49300
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Kyrgyz Republic               : <dendro::node>
#>  ..  ..@ index   : int 75
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "KGZ"
#>  .. .. .. $ population: num 5431747
#>  .. .. .. $ GNI       : int 1250
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Lao PDR                       : <dendro::node>
#>  ..  ..@ index   : int 76
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LAO"
#>  .. .. .. $ population: num 6834942
#>  .. .. .. $ GNI       : int 1660
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Lebanon                       : <dendro::node>
#>  ..  ..@ index   : int 77
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LBN"
#>  .. .. .. $ population: num 4017095
#>  .. .. .. $ GNI       : int 10030
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Macao SAR, China              : <dendro::node>
#>  ..  ..@ index   : int 78
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MAC"
#>  .. .. .. $ population: num 559846
#>  .. .. .. $ GNI       : int 76270
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Malaysia                      : <dendro::node>
#>  ..  ..@ index   : int 79
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MYS"
#>  .. .. .. $ population: num 25715819
#>  .. .. .. $ GNI       : int 11120
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Mongolia                      : <dendro::node>
#>  ..  ..@ index   : int 80
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MNG"
#>  .. .. .. $ population: num 3041142
#>  .. .. .. $ GNI       : int 4280
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Myanmar                       : <dendro::node>
#>  ..  ..@ index   : int 81
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "MMR"
#>  .. .. .. $ population: num 48137741
#>  .. .. .. $ GNI       : int 1270
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Nepal                         : <dendro::node>
#>  ..  ..@ index   : int 82
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "NPL"
#>  .. .. .. $ population: num 28563377
#>  .. .. .. $ GNI       : int 730
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Oman                          : <dendro::node>
#>  ..  ..@ index   : int 83
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "OMN"
#>  .. .. .. $ population: num 3418085
#>  .. .. .. $ GNI       : int 16870
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Pakistan                      : <dendro::node>
#>  ..  ..@ index   : int 84
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "PAK"
#>  .. .. .. $ population: num 1.76e+08
#>  .. .. .. $ GNI       : int 1400
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Philippines                   : <dendro::node>
#>  ..  ..@ index   : int 85
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "PHL"
#>  .. .. .. $ population: num 9.8e+07
#>  .. .. .. $ GNI       : int 3500
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Qatar                         : <dendro::node>
#>  ..  ..@ index   : int 86
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "QAT"
#>  .. .. .. $ population: num 833285
#>  .. .. .. $ GNI       : int 92200
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Saudi Arabia                  : <dendro::node>
#>  ..  ..@ index   : int 87
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SAU"
#>  .. .. .. $ population: num 28686633
#>  .. .. .. $ GNI       : int 25140
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Singapore                     : <dendro::node>
#>  ..  ..@ index   : int 88
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "SGP"
#>  .. .. .. $ population: num 4657542
#>  .. .. .. $ GNI       : int 55150
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Sri Lanka                     : <dendro::node>
#>  ..  ..@ index   : int 89
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "LKA"
#>  .. .. .. $ population: num 21324791
#>  .. .. .. $ GNI       : int 3460
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Tajikistan                    : <dendro::node>
#>  ..  ..@ index   : int 90
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TJK"
#>  .. .. .. $ population: num 7349145
#>  .. .. .. $ GNI       : int 1080
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Thailand                      : <dendro::node>
#>  ..  ..@ index   : int 91
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "THA"
#>  .. .. .. $ population: num 65905410
#>  .. .. .. $ GNI       : int 5780
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Timor-Leste                   : <dendro::node>
#>  ..  ..@ index   : int 92
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TMP"
#>  .. .. .. $ population: num 1201542
#>  .. .. .. $ GNI       : int 2680
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Turkey                        : <dendro::node>
#>  ..  ..@ index   : int 93
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TUR"
#>  .. .. .. $ population: num 76805524
#>  .. .. .. $ GNI       : int 10830
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Turkmenistan                  : <dendro::node>
#>  ..  ..@ index   : int 94
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "TKM"
#>  .. .. .. $ population: num 4884887
#>  .. .. .. $ GNI       : int 8020
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ United Arab Emirates          : <dendro::node>
#>  ..  ..@ index   : int 95
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "ARE"
#>  .. .. .. $ population: num 4798491
#>  .. .. .. $ GNI       : int 44600
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Uzbekistan                    : <dendro::node>
#>  ..  ..@ index   : int 96
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "UZB"
#>  .. .. .. $ population: num 27606007
#>  .. .. .. $ GNI       : int 2090
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Vietnam                       : <dendro::node>
#>  ..  ..@ index   : int 97
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "VNM"
#>  .. .. .. $ population: num 8.7e+07
#>  .. .. .. $ GNI       : int 1890
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ West Bank and Gaza            : <dendro::node>
#>  ..  ..@ index   : int 98
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "WBG"
#>  .. .. .. $ population: num 4550368
#>  .. .. .. $ GNI       : int 3060
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  .. $ Yemen, Rep.                   : <dendro::node>
#>  ..  ..@ index   : int 99
#>  ..  ..@ value   :'data.frame':	1 obs. of  3 variables:
#>  .. .. .. $ iso3      : chr "YEM"
#>  .. .. .. $ population: num 23822783
#>  .. .. .. $ GNI       : int 1300
#>  ..  ..@ parent  : int 52
#>  ..  ..@ children: int(0) 
#>  ..  [list output truncated]